refactor: move mocket.socket to mocket.core.socket
