+++ mocket/__init__.py
@@ -1,4 +1,5 @@
 from mocket.async_mocket import async_mocketize
+from mocket.core.socket import MocketSocket
 from mocket.entry import MocketEntry
 from mocket.mocket import Mocket
 from mocket.mocketizer import Mocketizer, mocketize
@@ -7,14 +8,16 @@ from mocket.ssl.context import MocketSSLContext
 # NOTE this is here for backwards-compat to keep old import-paths working
 from mocket.ssl.context import MocketSSLContext as FakeSSLContext
 
-__all__ = (
-    "async_mocketize",
-    "mocketize",
+__all__ = [
     "Mocket",
     "MocketEntry",
-    "Mocketizer",
     "MocketSSLContext",
+    "MocketSocket",
+    "Mocketizer",
+    "async_mocketize",
+    "mocketize",
+    # NOTE this is here for backwards-compat to keep old import-paths working
     "FakeSSLContext",
-)
+]
 
 __version__ = "3.13.2"
