+++ mocket/ssl/context.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import Any
 
-from mocket.socket import MocketSocket
+from mocket.core.socket import MocketSocket
 from mocket.ssl.socket import MocketSSLSocket
